fix: improve derived generation navigation message in Timeline
Updated Timeline and TimelineRefactored to show a more helpful message when
clicking on derived/edited generations that aren't in the timeline.

Changes:
- Fetch full generation data (select '*' instead of just 'id')
- Show clear message: 'This is a derived generation. Switch to Shot Editor
  to view all edits of this image.'
- Increased toast duration to 4000ms so users can read it
- Added TODO comment to implement full external generation viewing in Timeline
  (like ShotImageManager has with useExternalGenerations)

Context:
Timeline is optimized for viewing positioned timeline images. Derived/edited
generations (from magic edit, inpaint, etc.) are not positioned in the timeline
and are best viewed in Shot Editor, which has the infrastructure to navigate
through external generations and derived contexts.